% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{AAPLStockMonthly}
\alias{AAPLStockMonthly}
\title{Closing Stock Price of Apple Inc.}
\format{A data frame with 409 observations on the following variables.
  \itemize{
    \item{\code{Month}} {months since January 1981}
    \item{\code{AdjClose}} {adjusted closing price of Apple Inc. stock}
  }}
\source{
Yahoo Finance
}
\usage{
data(AAPLStockMonthly)
}
\description{
A dataset containing the closing price of Apple Inc. Stock, from January 1981 to December 2014
}
\details{
closing price of Apple Inc. stock in U.S. dollars adjusted for dividends and splits at the
beginning of each month from January 1981 (month 1) through December 2014 (month 408)
}
\examples{
data(APPLStockMontly)

}
\references{
http://finance.yahoo.com/q/hp?s=AAPL&a=11&b=12&c=1980&d=06&e=2&f=2014&g=m
}
