% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{Abalone}
\alias{Abalone}
\title{Abalone Rings}
\source{
Abalone: Donated on 11/30/1995 to UC Irvine Machine Learning Repository.
Accessed on July 19, 2025.
}
\usage{
data(Abalone)
}
\description{
A data set containing the number of rings for 4177 abalone along
with their length, and live weight. The age of an abalone is the number of
rings + 1.5, but it is difficult to count the number of rings to determine
the age of an abalone.
}
\details{
A data frame with 1307 observations on the following variables.
  \itemize{
    \item{\code{Length}} (in millimeters)
    \item{\code{Weight}} (live weight in grams)
    \item{\code{Rings}} (number of rings)
  }
}
\examples{
data(Abalone)

}
\references{
https://archive.ics.uci.edu/dataset/1/abalone
}
