% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{AddidasProfit}
\alias{AddidasProfit}
\title{Profit of Addidas}
\source{
Gross profit of Adidas from 2000 to 2024 (in million euros). Accessed on
May 31, 2025.
}
\usage{
data(AddidasProfit)
}
\description{
A data set containing the gross profit of Addidas in millions
of euros from 2006 to 2024.
}
\details{
A data frame with 19 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Profit}} (in millions of euros)
  }
}
\examples{
data(AddidasProfit)

}
\references{
https://www.statista.com/statistics/268417/gross-profit-of-the-adidas-group-worldwide-since-2000/
}
