% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{Advertising}
\alias{Advertising}
\title{Sales Revenue based on Advertising Budget Breakdown}
\source{
Advertising Sales Data Set: Advertising Budget & Sales Prediction using
Regression. Accessed on July 18, 2025.
}
\usage{
data(Advertising)
}
\description{
A data set containing the sales revenue of various products in
millions of US dollars based on the TV, radio, and newspaper advertising
budget of the product in 1000's of US dollars.
}
\details{
A data frame with 200 observations on the following variables.
  \itemize{
    \item{\code{Sales}} (in millions of US dollars)
    \item{\code{TV}} (TV advertising budget in 1000's of US dollars)
    \item{\code{Radio}} (radio advertising budget in 1000's of US dollars)
    \item{\code{Newspaper}} (newspaper advertising budget in 1000's of US dollars)
  }
}
\examples{
data(Advertising)

}
\references{
https://www.kaggle.com/datasets/yasserh/advertising-sales-dataset
}
