% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{APCalculus2}
\alias{APCalculus2}
\title{Number of Students taking the AP Calculus Exam}
\format{A data frame with 38 observations on the following variables.
  \itemize{
    \item{\code{Year}} {year}
    \item{\code{Exams}} {number of AP exams taken in associated year}
  }}
\source{
Personal correspondence with Stephen M. Kokoska.
}
\usage{
data(APCalculus2)
}
\description{
A subset of the dataset \code{APCalculus} containing the number of students who have taken the AP Calculus exam from 1955 to 2015.
}
\details{
A subset of the MMAC dataset APCalculus containing only the data from 1973 to 2010.
}
\examples{
data(APCalculus2)

}
