% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{AutoMPG}
\alias{AutoMPG}
\title{Fuel Efficiency of Automobiles}
\source{
Data on predicting the MPG of 392 automobiles based on the number of
cylinders, engine displacement, engine horsepower, vehicle weight, acceleration,
and year. Accessed on August 3, 2025.
}
\usage{
data(AutoMPG)
}
\description{
A data set containing data on predicting the MPG of 392
automobiles based on the number of cylinders, engine displacement, engine
horsepower, vehicle weight, acceleration, and year.
}
\details{
A data frame with 392 observations on the following variables.
  \itemize{
    \item{\code{MPG}}
    \item{\code{Cylinders}} (number of cylinders in engine)
    \item{\code{Displacement}} (in cubic inches)
    \item{\code{Horsepower}}
    \item{\code{Weight}} (in pounds)
    \item{\code{Acceleration}} (in meters per second squared)
    \item{\code{Year}} (year of production; e.g. 70 is the year 1970)
  }
}
\examples{
data(AutoMPG)

}
\references{
https://archive.ics.uci.edu/dataset/9/auto+mpg
}
