% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{AverageCO2Levels}
\alias{AverageCO2Levels}
\title{Average Worldwide CO2 Levels}
\source{
Average carbon dioxide levels in the atmosphere worldwide from 1959 to 2023
(in parts per million). Accessed on October 1, 2024.
}
\usage{
data(AverageCO2Levels)
}
\description{
A data set containing the average worldwide carbon dioxide concentration
level in parts per million from 1993 to 2023.
}
\details{
A data frame with 32 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Emissions}} (parts per million)
  }
}
\examples{
data(AverageCO2Levels)

}
\references{
https://www.statista.com/statistics/1091926/atmospheric-concentration-of-co2-historic/
}
