% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{AvianFemur}
\alias{AvianFemur}
\title{Length of Avian Femurs}
\source{
Prange, H.D., Anderson, J.F., and Rahn, H. Scaling of Skeletal Mass to Body
Mass in Birds and Mammals. \emph{The American Naturalist}, Vol. 113, No. 1
(Jan. 1979), pp. 103-122. Accessed on Augut 1, 2025.
}
\usage{
data(AvianFemur)
}
\description{
A data set containing the length of avian femurs in millimeters as
a function of the body mass of the avian species in kilograms.
}
\details{
A data frame with 22 observations on the following variables.
  \itemize{
    \item{\code{Mass}} (in kilograms)
    \item{\code{Length}} (in millimeters)
  }
}
\examples{
data(AvianFemur)

}
\references{
https://www.jstor.org/stable/2459945
}
