% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{BlastData}
\alias{BlastData}
\title{Blast Radius of Trinity Test}
\source{
Taylor, G. “The Formation of a Blast Wave by a Very Intense Explosion. II. The Atomic Explosion
of 1945.” \emph{Proceedings of the Royal Society A: Mathematical, Physical and Engineering Sciences},
1950, 201: 175–86.
}
\usage{
data(BlastData)
}
\description{
A data set containing the blast radius (in meters) as a function of time (in seconds) for the Trinity
test conducted by the United States in White Sands, New Mexico in 1945.
}
\details{
A data frame with 23 observations on the following variables.
  \itemize{
    \item{\code{time}} (in seconds)
    \item{\code{radius}} (in meters)
  }
}
\examples{
data(BlastData)

}
