% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{BodyMassMetabolicRate}
\alias{BodyMassMetabolicRate}
\title{Closing Stock Price of Apple Inc.}
\source{
Hudson, L. N., Isaac, N. J. B., Reuman, D. C. (2013), The relationship between body mass and
field metabolic rate among individual birds and mammals. \emph{Journal of Ecology}, 82: 1009-1020.
doi: 10.1111/1365-2656.12086.
}
\usage{
data(BodyMassMetabolicRate)
}
\description{
A data set containing the field metablic rate of individual birds and mammals (measured in
kilojoules per day) as a function of its body mass (in kilograms)
}
\details{
A data frame with 1498 observations on the following variables.
  \itemize{
    \item{\code{Mass}} (in kilograms)
    \item{\code{Rate}} (field metabolic rate in kilojoules per day)
  }
}
\examples{
data(BodyMassMetabolicRate)

}
\references{
http://onlinelibrary.wiley.com/doi/10.1111/1365-2656.12086/suppinfo
}
