% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{CaliforniaHousing}
\alias{CaliforniaHousing}
\title{California Housing Data from 1990 US Census}
\source{
Data on California housing from the 1990 census. Accessed on August 3, 2025.
}
\usage{
data(CaliforniaHousing)
}
\description{
A data set containing data on housing prices in California from
the 1990 Census. Each data point is from one census block and contains data
on the median age, total number of rooms, total number of bedrooms, population,
number of households, median income, and median value of the houses within
that census block.
}
\details{
A data frame with 20433 observations on the following variables.
  \itemize{
    \item{\code{Age}} (median age of house within block)
    \item{\code{Rooms}} (total number of rooms within block)
    \item{\code{Bedrooms}} (total number of bedrooms within block)
    \item{\code{Population}} (total number of people residing within block)
    \item{\code{Households}} (total number of households residing within block)
    \item{\code{MedianIncome}} (median income within block)
    \item{\code{MedianValue}} (median home value within block)
  }
}
\examples{
data(CaliforniaHousing)

}
\references{
https://www.dcc.fc.up.pt/~ltorgo/Regression/cal_housing.html
}
