% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{CarSales}
\alias{CarSales}
\title{Car Sales}
\source{
Cars for Sale: Comprehensive Dataset of Cars for Sale with Detailed Features,
Pricing, accident. Accessed on July 18, 2025.
}
\usage{
data(CarSales)
}
\description{
A data set containing the price, year, and mileage of 23000 cars
for sale.
}
\details{
A data frame with 23000 observations on the following variables.
  \itemize{
    \item{\code{Year}} (year car was manufactured)
    \item{\code{Mileage}} (number of miles on car)
    \item{\code{Price}} (price of car in US dollars)
  }
}
\examples{
data(CarSales)

}
\references{
https://www.kaggle.com/datasets/benjnb/cars-for-sale
}
