% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{CellularSubscriptions}
\alias{CellularSubscriptions}
\title{Cellular Subscriptions Worldwide}
\source{
Mobile cellular subscriptions (per 100 people). Accessed on October 2, 2024.
}
\usage{
data(CellularSubscriptions)
}
\description{
A data set containing the worldwide number of mobile cellular subscriptions
per 100 people from 1980 to 2008.
}
\details{
A data frame with 29 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Subscriptions}} (per 100 people)
  }
}
\examples{
data(CellularSubscriptions)

}
\references{
https://data.worldbank.org/indicator/IT.CEL.SETS.P2?end=2023&start=1960&view=chart
}
