% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{CellularSubscriptions2}
\alias{CellularSubscriptions2}
\title{Worldwide Mobile Cellular Subscriptions}
\source{
Mobile cellular subscriptions (per 100 people). Accessed on October 2, 2024.
}
\usage{
data(CellularSubscriptions2)
}
\description{
A data set containing the worldwide number of mobile cellular
subscriptions per 100 people from 1980 to 2023.
}
\details{
A data frame with 44 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Subscriptions}} (per 100 people)
  }
}
\examples{
data(CellularSubscriptions2)

}
\references{
https://data.worldbank.org/indicator/IT.CEL.SETS.P2
}
