% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{ChildLabor}
\alias{ChildLabor}
\title{Child Labor Law Violations}
\source{
Child Labor. Accessed on June 24, 2025.
}
\usage{
data(ChildLabor)
}
\description{
A data set containing the number of minors employed in the United
States in violation of child labor laws from 1971 to 2024.
}
\details{
A data frame with 28 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Violations}}
  }
}
\examples{
data(ChildLabor)

}
\references{
https://www.dol.gov/agencies/whd/data/charts/child-labor
}
