% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{CHIPExpenditures}
\alias{CHIPExpenditures}
\title{CHIP Consumption Expenditures}
\source{
Children's health insurance program (CHIP) health consumption expenditure
from 2000-2022 (in million U.S. dollars). Accessed on March 11, 2025.
}
\usage{
data(CHIPExpenditures)
}
\description{
A data set containing the Children's Health Insurance Program (CHIP)
consumption expenditures from 2005 to 2022 in millions of US dollars.
}
\details{
A data frame with 18 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Expenditure}} (in millions of US dollars)
  }
}
\examples{
data(CHIPExpenditures)

}
\references{
https://www.statista.com/statistics/245384/chip-personal-health-care-expenditure/
}
