% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{CigaretteSales}
\alias{CigaretteSales}
\title{United States Cigarette Sales}
\source{
Federal Trade Commission Cigarette Report for 2022. Accessed on June 5, 2025.
}
\usage{
data(CigaretteSales)
}
\description{
A data set containing the number of individual cigarettes sold
in the United States in billions each year from 1981 to 2022.
}
\details{
A data frame with 42 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Sales}} (in billions of individual cigarettes)
  }
}
\examples{
data(CigaretteSales)

}
\references{
https://www.ftc.gov/system/files/ftc_gov/pdf/2022-Cigarette-Report.pdf
}
