% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{DaylengthFairbanks}
\alias{DaylengthFairbanks}
\title{Daylength in Fairbanks, Alaska}
\source{
Fairbanks, Alaska, USA — Sunrise, Sunset, and Daylength, August 2025.
Accessed on August 4, 2025.
}
\usage{
data(DaylengthFairbanks)
}
\description{
A data set containing the daylength in Fairbanks, Alaska in hours
from August 2020 to August 2025.
}
\details{
A data frame with 61 observations on the following variables.
  \itemize{
    \item{\code{Month}} (since August 2020)
    \item{\code{Daylength}}  (in hours)
  }
}
\examples{
data(DaylengthFairbanks)

}
\references{
https://www.timeanddate.com/sun/usa/fairbanks
}
