% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{DJIA}
\alias{DJIA}
\title{Dow Jones Industrial Average Closing Value}
\source{
Closing stock market value of the Dow Jones Industrial Average at the end of
each quarter from June 30, 1896 through September 30, 2024 from
``^DJI --- Nasdaq Composite --- U.S. --- Stooq.'' Accessed on October 1, 2024.
}
\usage{
data(DJIA)
}
\description{
A data set containing the closing stock market value of the Dow Jones
Industrial Average on the last day of each quarter from June 30, 1896
(quarter 1) through September 30, 2024 (quarter 515).
}
\details{
A data frame with 515 observations on the following variables.
  \itemize{
    \item{\code{Quarter}}
    \item{\code{Close}}
  }
}
\examples{
data(DJIA)

}
\references{
http://stooq.com/q/d/?s=^dji&i=q.
}
