% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{DuluthDayLength}
\alias{DuluthDayLength}
\title{Day Length in Duluth, MN}
\source{
Duluth, Minnesota, USA -- Sunrise, Sunset, and Daylength, October 2024.
Accessed on October 24, 2024.
}
\usage{
data(DuluthDayLength)
}
\description{
A data set containing the day length in Duluth, MN measures in
hours on the first day of each month from October 2022 to October 2024.
}
\details{
A data frame with 25 observations on the following variables.
  \itemize{
    \item{\code{Month}} (since October 2022)
    \item{\code{Length}} (in hours)
  }
}
\examples{
data(DuluthDayLength)

}
\references{
https://www.timeanddate.com/sun/usa/duluth
}
