% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{DunkinDonuts}
\alias{DunkinDonuts}
\title{Dunkin' Donuts Stores Worldwide}
\source{
Number of Dunkin' Donuts stores worldwide from 2007 to 2019, by region.
Accessed on September 19, 2024.
}
\usage{
data(DunkinDonuts)
}
\description{
A data set containing the number of Dunkin' Donuts stores worldwide from
2007 to 2019.
}
\details{
A data frame with 13 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Stores}}
  }
}
\examples{
data(DunkinDonuts)

}
\references{
https://www.statista.com/statistics/291462/distribution-points-dunkin-donuts/
}
