% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{ElectricBill}
\alias{ElectricBill}
\title{Electric Bill}
\source{
Subset of Utilities in the \code{mosaic} package.
}
\usage{
data(ElectricBill)
}
\description{
A data set containing the electric bill (in US dollars) of a single-family home in Minnesota from
2000-2003
}
\details{
A data frame with 37 observations on the following variables.
  \itemize{
    \item{\code{Month}} (since January 2000)
    \item{\code{ElecBill}} (in US dollars)
  }
}
\examples{
data(ElectricBill)

}
