% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{ElectricityDemand}
\alias{ElectricityDemand}
\title{United States Electricity Demand}
\source{
U.S. electricity overview (demand, forecast demand, net generation, and total
interchange) 10/1/2024 - 10/5/24, Eastern Time. Accessed on October 24, 2024.
}
\usage{
data(ElectricityDemand)
}
\description{
A data set containing the United States electricity demand in
megawatt hours from October 1, 2024 to October 5, 2024 as a function of time
in hours.
}
\details{
A data frame with 120 observations on the following variables.
  \itemize{
    \item{\code{Time}} (hours since October 1, 2024)
    \item{\code{Length}} (in megawatt hours)
  }
}
\examples{
data(ElectricityDemand)

}
\references{
https://www.eia.gov/electricity/gridmonitor/dashboard/custom/pending
}
