% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{EngineBore}
\alias{EngineBore}
\title{Engine Bore Diameter}
\source{
Bore diameter as a function of engine mass from pages 60--61. McMahon,
Thomas A., and John Tyler Bonner. \emph{On Size and Life}. New York:
Scientific American Library, 1983.
}
\usage{
data(EngineBore)
}
\description{
A data set containing the bore diameter (i.e. the diameter of
the cylinder) in millimeters as a function of the mass of various engines
in kilograms
}
\details{
A data frame with 39 observations on the following variables.
  \itemize{
    \item{\code{Mass}} (in kilograms)
    \item{\code{BoreLength}} (in millimeters
  }
}
\examples{
data(EngineBore)

}
