% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{EngineRPM}
\alias{EngineRPM}
\title{RPM of Different Engines}
\source{
Revolutions per minute of engines as a function of engine mass from pages 60-61. McMahon,
Thomas A., and John Tyler Bonner. \emph{On Size and Life}. New York: Scientific American
Library, 1983.
}
\usage{
data(EngineRPM)
}
\description{
A data set containing the revolutions per minute (RPM) and mass of different engines.
}
\details{
A data frame with 39 observations on the following variables.
  \itemize{
    \item{\code{Mass}} (in kilograms)
    \item{\code{RPM}}
  }
}
\examples{
data(EngineRPM)

}
