% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{EVMarketShare}
\alias{EVMarketShare}
\title{Worldwide Electric Vehicle Market Share}
\source{
Global market share of electric vehicles within passenger vehicle sales
between 2015 and 2023, with a forecast through 2030. Accessed on June 18, 2025.
}
\usage{
data(EVMarketShare)
}
\description{
A data set containing the worldwide market share of electric
vehicles as a percentage of passenger vehicle sales from 2015 to 2023.
}
\details{
A data frame with 9 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Share}}
  }
}
\examples{
data(EVMarketShare)

}
\references{
https://www.statista.com/statistics/1371599/global-ev-market-share/
}
