% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{FacebookUsers}
\alias{FacebookUsers}
\title{Facebook Users}
\source{
Number of Facebook users in millions from 2009 through 2012 from “Number of active users at Facebook
over the years — Yahoo News” and “Facebook: number of active users 2015 | Statistics.”
Accessed on June 22, 2015.
}
\usage{
data(FacebookUsers)
}
\description{
A data set containing the number of Facebook users in millions of people from 2009
through 2012.
}
\details{
A data frame with 38 observations on the following variables.
  \itemize{
    \item{\code{Months}} (since January 2009)
    \item{\code{Users}} (in millions)
  }
}
\examples{
data(FacebookUsers)

}
\references{
http://news.yahoo.com/number-active-users-facebook-over-230449748.html
http://www.statista.com/statistics/264810/number-of-monthly-active-facebook-users-worldwide/
}
