% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{FiddlerCrabs}
\alias{FiddlerCrabs}
\title{Palm Length of Fiddler Crabs}
\source{
Miller, D.C. Growth in Uca, 1. Ontogeny of Asymmetry in Uca pugilator (Bosc)
(Decapoda, Ocypodidae). \emph{Crustaceana}, Vol. 24, No. 1 (Jan., 1973), pp.
119-131. Accessed on August 1, 2025.
}
\usage{
data(FiddlerCrabs)
}
\description{
A data set containing the palm length (chela size) of growing
male fiddler crabs as a function of the body size (carapace breadth), both
measured in millimeters.
}
\details{
A data frame with 19 observations on the following variables.
  \itemize{
    \item{\code{BodySize}} (in millimeters)
    \item{\code{PalmLength}}  (in millimeters)
  }
}
\examples{
data(FiddlerCrabs)

}
\references{
https://www.jstor.org/stable/20101959
}
