% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{FishWeights}
\alias{FishWeights}
\title{Weight of Bream}
\source{
FishMarket: Estimate the weight of fish based on its species and the physical
measurements. Accessed on August 3, 2025
}
\usage{
data(FishWeights)
}
\description{
Weight of 56 bream fish in grams as a function of their vertical
length (in cm), diagonal length (in cm), cross length (in cm), height (in cm),
and diagonal width (in cm).
}
\details{
A data frame with 56 observations on the following variables.
  \itemize{
    \item{\code{Weight}} (in grams)
    \item{\code{Vertial}} (vertical length in centimeters)
    \item{\code{Diagonal}} (diagonal length in centimeters)
    \item{\code{Cross}} (cross length in centimeters)
    \item{\code{Height}} (in centimeters)
    \item{\code{Width}} (diagonal width in centimeters)
  }
}
\examples{
data(FishWeights)

}
\references{
https://www.kaggle.com/datasets/vipullrathod/fish-market
}
