% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{FlightsWorldwide}
\alias{FlightsWorldwide}
\title{Worldwide Flights Performed}
\source{
Number of flights performed by the global airline industry from 2004 to 2024,
with a forecast for 2025 (in millions). Accessed on August 2, 2025.
}
\usage{
data(FlightsWorldwide)
}
\description{
A data set containing the worldwide number of flights performed
by the global airline industry in millions from 2004 to 2019.
}
\details{
A data frame with 16 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Flights}}  (in millions)
  }
}
\examples{
data(FlightsWorldwide)

}
\references{
https://www.statista.com/statistics/564769/airline-industry-number-of-flights/
}
