% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{FordMarketVolume2}
\alias{FordMarketVolume2}
\title{Ford Motors Market Volume}
\source{
Ford Motor Company (F) stock market value quarterly in U.S. dollars. Accessed on July 7, 2014
}
\usage{
data(FordMarketVolume2)
}
\description{
A data set containing the volume of Ford Motor Company stock shares traded per
quarter since January 1, 2007.
}
\details{
A data frame with 20 observations on the following variables.
  \itemize{
    \item{\code{Date}} (quarter since January 1, 2007)
    \item{\code{Volume}} (number of shares of Ford Motor Company stock shares)
  }
}
\examples{
data(FordMarketVolume1)

}
\references{
http://stooq.com/q/d/?s=f.us
}
