% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{GaltonMale}
\alias{GaltonMale}
\title{Height of Children}
\source{
Hanley, J.A. (2004). ``Transmuting" women into men: Galton's family data on
human stature. The American Statistician, vol. 58, no. 3, pp. 237-243.
Accessed on May 21, 2025.
}
\usage{
data(GaltonMale)
}
\description{
A data set containing data on the height of 465 male children and the
height of both their mother and father.
}
\details{
A data frame with 465 observations on the following variables.
  \itemize{
    \item{\code{father}} (height of father in inches)
    \item{\code{mother}} (height of mother in inches)
    \item{\code{height}} (height of child in inches)
  }
}
\examples{
data(GaltonMale)

}
\references{
https://www.medicine.mcgill.ca/epidemiology/hanley/galton/
}
