% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{GenderRatio1}
\alias{GenderRatio1}
\title{Gender Ratio in World Population}
\format{A data frame with 9 observations on the following variables.
  \itemize{
    \item{\code{Year}} {}
    \item{\code{Ratio}} {number of males per 100 females}
  }}
\source{
The global gender ratio based on the number of males per 100 females by year from
“World Population Prospects, the 2012 Revision” by the United Nations Department of
Economic and Social Affairs. Accessed on June 25, 2014.
}
\usage{
data(GenderRatio1)
}
\description{
A subset of the dataset containing the number of males per 100 females in the word population.
}
\details{
This dataset contains the number of males per 100 females in the world population as a function of the year
}
\examples{
data(GenderRatio1)

}
\references{
http://esa.un.org/unpd/wpp/Excel-Data/population.htm
}
