% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{GlobalCO2Emissions}
\alias{GlobalCO2Emissions}
\title{Global CO2 Emissions}
\source{
Annual carbon dioxide emissions worldwide from 1940 to 2023 (in billion
metric tons). Accessed on September 26, 2024.
}
\usage{
data(GlobalCO2Emissions)
}
\description{
A data set containing the annual worldwide carbon dioxide emissions in billion
metric tons from 1940 to 2023.
}
\details{
A data frame with 84 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Emissions}} (in billion metric tons)
  }
}
\examples{
data(GlobalCO2Emissions)

}
\references{
https://www.statista.com/statistics/276629/global-co2-emissions/
}
