% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{Happiness}
\alias{Happiness}
\title{Happiness in Southeastern Asian and Middle Eastern Countries}
\source{
SA-ME Happiness Index: Education, Income & Life Satisfaction Metrics.
Accessed on July 18, 2025.
}
\usage{
data(Happiness)
}
\description{
A data set containing a happiness score (between 0 and 10) from
the 2023 World Happiness Report from 12 Southeastern Asian and Middle Eastern
countries along with the median income in United States dollars and the poverty
rate of the county.
}
\details{
A data frame with 12 observations on the following variables.
  \itemize{
    \item{\code{Happiness}} (score between 0 and 10)
    \item{\code{Income}} (median income in US dollars)
    \item{\code{Poverty}} (poverty rate)
  }
}
\examples{
data(Happiness)

}
\references{
https://www.kaggle.com/datasets/towhid121/sa-me-happiness-index
}
