% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{Hawaii}
\alias{Hawaii}
\title{Tidal Depths in Pearl Harbor, Hawaii}
\format{A data frame with 31 observations on the following variables.
  \itemize{
    \item{\code{time}} {in hours}
    \item{\code{water}} {tide depth in feet relative to the MLLW}
  }}
\source{
Tidal measurements in Pearl Harbor, Hawaii based on data sets from Project Mosaic. Accessed on
August 19, 2015.
}
\usage{
data(Hawaii)
}
\description{
A dataset containing the depth of the tide in feet relative to the MLLW (mean lower low water mark)
in Pearl Harbor, Hawaii as a function of time measured in hours Hr.
}
\details{
This dataset contains the depth of the tide in feet relative to the MLLW (mean lower low water
mark) in Pearl Harbor, Hawaii as a function of time measured in hours Hr.
}
\examples{
data(Hawaii)

}
\references{
http://www.mosaic-web.org
}
