% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{HispanicPopulation}
\alias{HispanicPopulation}
\title{Latino's Living in the United States}
\format{A data frame with 5 observations on the following variables.
  \itemize{
    \item{\code{Year}} {}
    \item{\code{People}} {number of Latino's (in millions)}
  }}
\source{
The U.S. Hispanic population has increased sixfold since 1970. Accessed on June 16, 2014.
}
\usage{
data(HispanicPopulation)
}
\description{
A dataset containing the number of Latino's living in the United States as a function of the year.
}
\details{
This dataset contains the number of Latino's living in the United States as a function of the year.
}
\examples{
data(HispanicPopulation)

}
\references{
http://www.pewresearch.org/fact-tank/2014/02/26/the-u-s-hispanic-population-has-increased-sixfold-since-1970/
}
