% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{HSGradsInCollege}
\alias{HSGradsInCollege}
\title{High School Graduates in College}
\source{
Percent of high school graduates to enroll in a two-year or four-year college from National Center
for Education Statistics. Accessed on July 2, 2014.
}
\usage{
data(HSGradsInCollege)
}
\description{
A data set containing the percent of each year’s high school
graduates to enroll in either a two-year or four-year college from 1972 to 2012.
}
\details{
A data frame with 41 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Percent}} (of high schol graduates)
  }
}
\examples{
data(HSGradsInCollege)

}
\references{
http://nces.ed.gov/programs/digest/d13/tables/dt13302.10.asp
}
