% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{JuniperTrees}
\alias{JuniperTrees}
\title{Biomass of OneSeed Juniper trees}
\source{
Boschetti Fabio , Sauer Katherine J. , Litvak Marcy , Anderson Karen ,
Brazier Richard E., Allometric Relationships for Predicting Aboveground
Biomass and Sapwood Area of Oneseed Juniper (Juniperus monosperma) Trees,
Frontiers in Plant Science, Volume 11, 2020, doi:10.3389/fpls.2020.00094.
Accessed on October 9, 2024.
}
\usage{
data(JuniperTrees)
}
\description{
A data set containing the above ground biomass of Oneseed juniper trees in
kilograms as a function of the maximum height of the tree is meters.
}
\details{
A data frame with 18 observations on the following variables.
  \itemize{
    \item{\code{Height}} (in meters)
    \item{\code{BioMass}} (in kilograms)
  }
}
\examples{
data(JuniperTrees)

}
\references{
https://www.frontiersin.org/journals/plant-science/articles/10.3389/fpls.2020.00094
}
