% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{KeyWestTides}
\alias{KeyWestTides}
\title{KeyWestTides}
\source{
NOAA/NOS/CO-OPS Observed Water Levels at 8724580, Key West, FL from
2024/09/09 00:00 GMT to 2024/09/12 23:59 GMT. Accessed on October 23, 2024
}
\usage{
data(KeyWestTides)
}
\description{
A data set containing the tidal depth relative to the mean lower
low water (MLLW) level in feet as a function of time in minutes in Key West,
Florida.
}
\details{
A data frame with 960 observations on the following variables.
  \itemize{
    \item{\code{Time}} (in minutes)
    \item{\code{Depth}} (in feet)
  }
}
\examples{
data(KeyWestTides)

}
\references{
https://tidesandcurrents.noaa.gov/waterlevels.html?id=8724580
}
