% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{LifeExpectancyPhysicians}
\alias{LifeExpectancyPhysicians}
\title{Life Expectancy in Different Countries}
\format{A data frame with 175 observations on the following variables.
  \itemize{
    \item{\code{LifeExpectancy}} {life expectancy in years}
    \item{\code{Physicians}} {number of physicians per 1000 people}
  }}
\source{
Number of physicians per 1000 people as a function of average life expectancy in different
countries in 2010 from “Physicians (per 1,000 people) | Data | Table.”
}
\usage{
data(LifeExpectancyPhysicians)
}
\description{
A dataset containing the number of physicians per 1000 people as a function of average
life expectancy in different countries in 2010.
}
\details{
This dataset contains the number of physicians per 1000 people as a function of average
life expectancy in different countries in 2010.
}
\examples{
data(LifeExpectancyPhysicians)

}
\references{
http://data.worldbank.org/indicator/SH.MED.PHYS.ZS
}
