% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{MammalEnergy}
\alias{MammalEnergy}
\title{Metablic Energy of Mammals}
\source{
Atanasov AT. The linear allometric relationship between total metabolic
energy per life span and body mass of mammals. Biosystems. 2007 Jul-Aug;
90(1):224-33. doi: 10.1016/j.biosystems.2006.08.006. Epub 2006 Aug 23.
PMID: 17030408. Accessed on October 10, 2024.
}
\usage{
data(MammalEnergy)
}
\description{
A data set containing the total lifetime metabolic energy in
kilojoules for various mammals as a function of their body mass in kilograms
}
\details{
A data frame with 93 observations on the following variables.
  \itemize{
    \item{\code{Mass}} (in kilograms)
    \item{\code{Energy}} (kilojoules)
  }
}
\examples{
data(MammalEnergy)

}
\references{
https://www.sciencedirect.com/science/article/abs/pii/S0303264706001481?via\%3Dihub
}
