% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{MaunaLoaCO2}
\alias{MaunaLoaCO2}
\title{Atmospheric Carbon Dioxide from Mauna Loa}
\source{
Atmospheric carbon dioxide from Mauna Loa in ppmv (parts per million by volume) as a function of
years from 1958 to 2008 from “Atmospheric Carbon Dioxide Record from Mauna Loa.” Accessed on
August 19, 2015.
}
\usage{
data(MaunaLoaCO2)
}
\description{
A data set containing the atmospheric carbon dioxide from Mauna Loa in ppmv (or parts per million
by volume) as a function of years from 1958 to 2008.
}
\details{
A data frame with 49 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Average}} (in ppmv)
  }
}
\examples{
data(MaunaLoaCO2)

}
\references{
http://cdiac.ornl.gov/trends/co2/sio-mlo.html
}
