% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{MediterraneanSeaTemp}
\alias{MediterraneanSeaTemp}
\title{Temperature of Mediterranean Sea}
\source{
Sea surface temperature anomalies in the Mediterranean Sea from 1994 to 2023.
Accessed on June 13, 2025.
}
\usage{
data(MediterraneanSeaTemp)
}
\description{
A data set containing the sea surface temperature of the
Mediterranean Sea in degrees Celsius from 1996 to 2023 relative to the
1991-2020 average.
}
\details{
A data frame with 28 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Temperature}} (in degrees Celsius relative to the 1991-2020 average)
  }
}
\examples{
data(MediterraneanSeaTemp)

}
\references{
https://www.statista.com/statistics/1463322/sea-surface-temperature-anomalies-mediterranean-sea/
}
