% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMAC-package.R
\docType{package}
\name{MMAC-package}
\alias{MMAC}
\alias{MMAC-package}
\title{MMAC: Data for Mathematical Modeling and Applied Calculus}
\description{
Contains the data sets for the first and second editions of the textbook "Mathematical Modeling and Applied Calculus" by Joel Kilty and Alex M. McAllister. The first edition of the book was published by Oxford University Press in 2018 with ISBN-13: 978-019882472. The second edition is expected to be published in January 2027.
}
\author{
\strong{Maintainer}: Joel Kilty \email{joel.kilty@centre.edu}

Authors:
\itemize{
  \item McAllister Alex \email{alex.mcallister@centre.edu}
}

}
\keyword{internal}
