% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespaceDoc.R
\docType{package}
\name{MMAC}
\alias{MMAC}
\alias{MMAC-package}
\title{Data for Mathematical Modeling and Applied Calculus}
\description{
Data for Mathematical Modeling and Applied Calculus
}
\details{
Data sets from the textbook Mathematical Modeling and Applied Calculus as well as all of the R
pacakges developed by Project MOSAIC needed for someone studying from this book.
}
\author{
Joel Kilty (\email{joel.kilty@centre.edu}), Alex M. McAllister (\email{alex.mcallister@centre.edu})
}
\keyword{package}
