% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{MonthlyUnemployment}
\alias{MonthlyUnemployment}
\title{US Unemployment Rate}
\source{
U.S. monthly unemployment rate from January 2010 to December 2014 from “Bureau of Labor Statistics Data.”
Accessed on June 22, 2015.
}
\usage{
data(MonthlyUnemployment)
}
\description{
A data set containing the United States monthly unemployment rate from January 2010 to December 2014.
}
\details{
A data frame with 60 observations on the following variables.
  \itemize{
    \item{\code{Months}} (since January 2010)
    \item{\code{Rate}} (monthly unemployment rate)
  }
}
\examples{
data(MonthlyUnemployment)

}
\references{
http://data.bls.gov/timeseries/LNS14000000
}
