% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{Mortgage15YrAnnual}
\alias{Mortgage15YrAnnual}
\title{15 Year Annual Mortgage Rates}
\source{
Interest rates on 15-year, fixed-rate conventional home mortgages annually from 1992 to 2014 from
“Mortgage Interest Rates History.” Accessed on June 22, 2015.
}
\usage{
data(Mortgage15YrAnnual)
}
\description{
A data set containing interest rates on 15-year, fixed-rate conventional home mortgages annually from
1992 to 2014.
}
\details{
A data frame with 23 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Rate}} (interest rate on 15-year, fixed-rate conventional home mortgage)
  }
}
\examples{
data(Mortgage15YrAnnual)

}
\references{
http://www.fedprimerate.com/mortgagerates.htm
}
