% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{Mortgage30YrAnnual}
\alias{Mortgage30YrAnnual}
\title{30 Year Annual Mortgage Rates}
\source{
Average interest rate for conventional 30-year mortgages each year from 1981 to 2012 from
“Primary Mortgage Market Survey Archives — 30 Year Fixed Rate Mortgages — Freddie Mac.”
Accessed on July 7, 2015.
}
\usage{
data(Mortgage30YrAnnual)
}
\description{
A data set containing the average interest rate for conventional 30-year mortgages each year
from 1981 to 2012.
}
\details{
A data frame with 32 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Rate}} (average interest rate on conventional 30-year mortgage)
  }
}
\examples{
data(Mortgage30YrAnnual)

}
\references{
http://www.freddiemac.com/pmms/pmms30.htm
}
