% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{Mortgage30YrMonthly1}
\alias{Mortgage30YrMonthly1}
\title{30 Year Annual Mortgage Rates}
\source{
Average interest rate for conventional 30-year mortgages each year from 1981 to 2012 from
“Primary Mortgage Market Survey Archives — 30 Year Fixed Rate Mortgages — Freddie Mac.”
Accessed on July 7, 2015.
}
\usage{
data(Mortgage30YrMonthly1)
}
\description{
A subset of the data set containing the interest rate on a 30-year fixed-rate conventional home mortgage.
}
\details{
A data frame with 265 observations on the following variables.
  \itemize{
    \item{\code{Month}}
    \item{\code{Rate}} (average interest rate on conventional 30-year mortgage)
  }
}
\examples{
data(Mortgage30YrMonthly1)

}
\references{
http://www.freddiemac.com/pmms/pmms30.htm
}
