% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{MusselsMetabolic}
\alias{MusselsMetabolic}
\title{Metabolic Rate of Mussels}
\source{
Ibarrola I, Arranz K, Markaide P, Navarro E (2022) Metabolic size scaling
reflects growth performance effects on age-size relationships in mussels
(Mytilus galloprovincialis). PLoS ONE 17(9): e0268053. Accessed on October 8, 2024.
}
\usage{
data(MusselsMetabolic)
}
\description{
A data set containing the standard metabolic rate in mL/hr of mussels
as a function of their live weight in grams
}
\details{
A data frame with 47 observations on the following variables.
  \itemize{
    \item{\code{LW}} (in grams)
    \item{\code{SMR}} (in mL/hr)
  }
}
\examples{
data(MusselsMetabolic)

}
\references{
https://doi.org/10.1371/journal.pone.0268053.
}
