% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{NASDAQ}
\alias{NASDAQ}
\title{Closing NASDAQ Market Value}
\source{
Closing NASDAQ stock market value in U.S. dollars at the end of each quarter
from March 1938 through September 2024 from ``Stooq:Historical data: Nasdaq
Composite - U.S. '' Accessed on September 27, 2024.
}
\usage{
data(NASDAQ)
}
\description{
A data set containing the closing NASDAQ stock market value in U.S. dollars at
the end of each quarter from March 1938 (Quarter=1) through September 2024
(Quarter=347).
}
\details{
A data frame with 347 observations on the following variables.
  \itemize{
    \item{\code{Quarter}}
    \item{\code{Close}}
  }
}
\examples{
data(NASDAQ)

}
\references{
https://stooq.com/q/d/?s=\%5Endq&c=0&i=q
}
