% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{NikeRevenue}
\alias{NikeRevenue}
\title{Revenue of Nike}
\source{
Revenue of Nike from 2005 to 2024 (in million U.S. dollars). Accessed on
June 18, 2025.
}
\usage{
data(NikeRevenue)
}
\description{
A data set containing the revenue of Nike in millions of US
dollars from 2010 to 2023.
}
\details{
A data frame with 14 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Revenue}}  (in millions of US dollars)
  }
}
\examples{
data(NikeRevenue)

}
\references{
https://www.statista.com/statistics/241683/nikes-sales-worldwide-since-2004/
}
