% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{NintendoSwitch}
\alias{NintendoSwitch}
\title{Nintendo Switch Sales}
\source{
Annual unit sales of the Nintendo Switch worldwide from 2017 to 2024.
Accessed on August 3, 2025.
}
\usage{
data(NintendoSwitch)
}
\description{
A data set containing the cumulative sales of the Nintendo Switch worldwide from 2017 to
2024.
}
\details{
A data frame with 8 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Sales}} (in millions of units)
  }
}
\examples{
data(NintendoSwitch)

}
\references{
https://www.statista.com/statistics/1085606/annual-unit-sales-nintendo-switch-worldwide/
}
