% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{NYCHousingPrices}
\alias{NYCHousingPrices}
\title{New York City Housing Prices}
\source{
Median sales price of housing in New York City from January 2019 to October
2024 (in U.S. dollars). Accessed on May 31, 2025.
}
\usage{
data(NYCHousingPrices)
}
\description{
A data set containing the contains the median sales price of
housing in New York City from October 2022 (Month=0) to October 2024 (Month=24).
}
\details{
A data frame with 25 observations on the following variables.
  \itemize{
    \item{\code{Month}} (since October 2022)
    \item{\code{Price}} (in US dollars)
  }
}
\examples{
data(NYCHousingPrices)

}
\references{
https://www.statista.com/statistics/1235655/median-sale-price-of-housing-new-york/
}
