% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{OilProductionAnnual2}
\alias{OilProductionAnnual2}
\title{Annual US Oil Production}
\source{
U.S. field production of crude oil in billions of barrels from the U.S. Energy Information
Administration from “U.S. Field Production of Crude Oil.” Accessed on July 3, 2014.
}
\usage{
data(OilProductionAnnual2)
}
\description{
A subset of the data set containing the number of barrels (in thousands) of crude oil produced per year
in the United States.
}
\details{
A data frame with 114 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Barrels}} (in thousands)
  }
}
\examples{
data(OilProductionAnnual2)

}
\references{
http://www.eia.gov/dnav/pet/hist/LeafHandler.ashx?n=PET&s=MCRFPUS1&f=M
}
