% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{OnlineAdvertising}
\alias{OnlineAdvertising}
\title{Online Advertising Revenue}
\source{
Online advertising revenue in the United States from 2000 to 2024 (in billion
U.S. dollars). Accessed on August 3, 2025.
}
\usage{
data(OnlineAdvertising)
}
\description{
A data set containing the online advertising revenue in the
United States from 2000 to 2024 in billion of U.S. dollars.
}
\details{
A data frame with 25 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Revenue}} (in billions of US dollars)
  }
}
\examples{
data(OnlineAdvertising)

}
\references{
https://www.statista.com/statistics/183816/us-online-advertising-revenue-since-2000/
}
