% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{OrganicFood}
\alias{OrganicFood}
\title{Organic Food Sales}
\source{
Organic food sales in the United States from 2005 to 2023 (in billion U.S.
dollars). Accessed on September 26, 2024.
}
\usage{
data(OrganicFood)
}
\description{
A data set containing the organic food sales in the United States in billions
of U.S. dollars from 2005 to 2023.
}
\details{
A data frame with 19 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Sales}} (billions of US dollars)
  }
}
\examples{
data(OrganicFood)

}
\references{
https://www.statista.com/statistics/196952/organic-food-sales-in-the-us-since-2000/
}
