% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{PaypalAccounts}
\alias{PaypalAccounts}
\title{Active Paypal Accounts}
\source{
Global user number of PayPal from 1st quarter 2010 to 2nd quarter 2024. Accessed
on October 2, 2024.
}
\usage{
data(PaypalAccounts)
}
\description{
A data set containing the global number of active Paypal accounts in millions
each quarter from the first quarter in 2010 (Quarter=1).
}
\details{
A data frame with 48 observations on the following variables.
  \itemize{
    \item{\code{Quarter}}
    \item{\code{Accounts}} (in millions)
  }
}
\examples{
data(PaypalAccounts)

}
\references{
https://www.statista.com/statistics/218493/paypals-total-active-registered-accounts-from-2010/
}
