% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{PaypalUsers}
\alias{PaypalUsers}
\title{Worldwide Number of Paypal Accounts}
\source{
Global user number of Paypal from 1st quarter 2010 to 1st quarter 2025.
Accessed on June 18, 2025.
}
\usage{
data(PaypalUsers)
}
\description{
A data set containing the worldwide number of Paypal accounts in
millions from the first quarter of 2010 (Quarter=1) to the first quarter of 2025.
}
\details{
A data frame with 61 observations on the following variables.
  \itemize{
    \item{\code{Quarter}} (since first Quarter of 2010)
    \item{\code{Accounts}}  (in millions)
  }
}
\examples{
data(PaypalUsers)

}
\references{
https://www.statista.com/statistics/218493/paypals-total-active-registered-accounts-from-2010/
}
