% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{PesticideConsumption}
\alias{PesticideConsumption}
\title{Pesticide Consumption}
\source{
Agricultural consumption of pesticides worldwide from 1990 to 2022
(in million metric tons). Accessed on September 26, 2024.
}
\usage{
data(PesticideConsumption)
}
\description{
A data set containing the worldwide agricultural consumption of pesticides
in million metric tons from 1990 to 2022.
}
\details{
A data frame with 33 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Pesticide}} (in million metric tons)
  }
}
\examples{
data(PesticideConsumption)

}
\references{
https://www.statista.com/statistics/1263077/global-pesticide-agricultural-use/
}
