% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{PollenCountLA}
\alias{PollenCountLA}
\title{Pollen Count in Los Angeles, CA}
\source{
Three-year average monthly pollen count in Los Angeles, California from “Historic Allergy Index
for 90001 | Pollen.com.” Accessed on June 25, 2014.
}
\usage{
data(PollenCountLA)
}
\description{
A data set containing the pollen count each day in Los Angeles from April to July of 2014.
}
\details{
A data frame with 30 observations on the following variables.
  \itemize{
    \item{\code{Date}} (since April 1, 2014)
    \item{\code{Count}} (Pollen count in Los Angeles, CA)
  }
}
\examples{
data(PollenCountLA)

}
\references{
http://www.pollen.com/allergy-trends.asp?PostalCode=90001
}
