% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{RoosterWeight}
\alias{RoosterWeight}
\title{Mean Rooster Weight}
\source{
Aggrey, S.E. Comparison of Three Nonlinear and Spline Regression Models for
Describing Chicken Growth Curves. \emph{Poultry Science}, Volume 81, Issue 12,
December 1, 2022, 1782-1788.
}
\usage{
data(RoosterWeight)
}
\description{
A data set containing the mean weight of roosters in grams as a
function of the age in days.
}
\details{
A data frame with 28 observations on the following variables.
  \itemize{
    \item{\code{Day}}
    \item{\code{Weight}}  (in grams)
  }
}
\examples{
data(RoosterWeight)

}
