% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{SATMathKentucky}
\alias{SATMathKentucky}
\title{SAT Math Scores in Kentucky}
\format{A data frame with 34 observations on the following variables.
  \itemize{
    \item{\code{Year}} {}
    \item{\code{Score}} {Average SAT math score in Kentucky}
  }}
\source{
Average SAT math score in Kentucky each year from 1980 to 2013 from College Board’s 2013 SAT
State Profile Report for Kentucky. Accessed on July 16, 2014.
}
\usage{
data(SATMathKentucky)
}
\description{
A dataset containing the average SAT math score in Kentucky each year from 1980 to 2013.
}
\details{
This dataset contains the average SAT math score in Kentucky each year from 1980 to 2013.
}
\examples{
data(SATMathKentucky)

}
\references{
http://media.collegeboard.com/digitalServices/pdf/research/2013/KY_13_03_03_01.pdf
}
