% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{SeaTemperature}
\alias{SeaTemperature}
\title{Average Global Sea Surface Temperature}
\source{
Average global sea surface temperature from ``Climate Change Indicators: Sea
Surface Temperature.'' Accessed on April 25, 2024.
}
\usage{
data(SeaTemperature)
}
\description{
A data set containing the average sea surface temperature from 1910 to 2020,
measured in degrees Fahrenheit from a 1971-2000 baseline average.
}
\details{
A data frame with 111 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Temperature}} (in degrees Fahrenheit)
  }
}
\examples{
data(SeaTemperature)

}
\references{
https://www.epa.gov/climate-indicators/climate-change-indicators-sea-surface-temperature
}
