% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{SmartphoneUsers}
\alias{SmartphoneUsers}
\title{Smartphone Users Worldwide}
\source{
Number of smartphone users worldwide from 2014 to 2029 (in millions).
Accessed on October 2, 2024.
}
\usage{
data(SmartphoneUsers)
}
\description{
A data set containing the worldwide number of smartphone users from 2014-2024
in millions.
}
\details{
A data frame with 11 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Users}} (in millions)
  }
}
\examples{
data(SmartphoneUsers)

}
\references{
https://www.statista.com/forecasts/1143723/smartphone-users-in-the-world
}
