% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{SolarCapacity}
\alias{SolarCapacity}
\title{Solar photovoltaic capacity in United States}
\source{
Cumulative solar PV capacity in the United States from 2000 to 2023
(in megawatts). Accessed on October 2, 2024.
}
\usage{
data(SolarCapacity)
}
\description{
A data set containing the cumulative solar photovoltaic capacity in the
United States from 2000 to 2023 measured in megawatts.
}
\details{
A data frame with 24 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Capacity}} (in megawatts)
  }
}
\examples{
data(SolarCapcity)

}
\references{
https://www.statista.com/statistics/232863/cumulative-solar-pv-capacity/
}
