% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{StJohnsTides}
\alias{StJohnsTides}
\title{Tides in St. Johns, Newfoundland}
\source{
Water levels at St. Johns relative to chart datum. St. Johns - 00905.
Accessed on August 2, 2025.
}
\usage{
data(StJohnsTides)
}
\description{
A data set containing the water levels in meters at St. Johns,
Newfoundland on the Bay of Fundy each hour from midnight July 13, 2025 to
midnight July 15, 2025.
}
\details{
A data frame with 672 observations on the following variables.
  \itemize{
    \item{\code{Hour}} (since midnight July 23, 2025)
    \item{\code{Level}}  (in meters)
  }
}
\examples{
data(StJohnsTides)

}
\references{
https://tides.gc.ca/en/stations/00905/2025-07-13
}
