% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{StudentDebt1}
\alias{StudentDebt1}
\title{Average Student Debt}
\source{
Average debt load in 2013 dollars of bachelor’s degree recipients attending U.S. public
colleges and universities who borrowed money to finance their education from “Average Cumulative
Debt Load of Bachelor’s Degree Recipients at Public Four-Year Institutions over Time — Trends
in Higher Education.” Accessed on July 5, 2015.
}
\usage{
data(StudentDebt1)
}
\description{
A subset of the data set containing the average cumulative debt of bachelor’s degree students enrolled in
public colleges and universities.
}
\details{
A data frame with 7 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Debt}} (Average cummulative debt)
  }
}
\examples{
data(StudentDebt1)

}
\references{
https://trends.collegeboard.org/student-aid/figures-tables/average-debt-levels-public-sector-bachelors-degree-recipients-over-time
}
