% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{StudentLoans}
\alias{StudentLoans}
\title{Student Loan Debt in United States}
\source{
Value of outstanding student loans in the United States from Q1 2006 to Q4
2024 (in billion U.S. dollars). Accessed on August 3, 2025.
}
\usage{
data(StudentLoans)
}
\description{
A data set containing the value of outstanding student loans in
the United States in billions of US dollars each quarter from the first quarter
of 2006 (Quarter=1) to the fourth quarter of 2024.
}
\details{
A data frame with 76 observations on the following variables.
  \itemize{
    \item{\code{Quarter}} (quarter since first quarter of 2006)
    \item{\code{Loans}} (in billions of US dollars)
  }
}
\examples{
data(StudentLoans)

}
\references{
https://www.statista.com/statistics/1077038/value-outstanding-student-loans-us/
}
