% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{SunflowerHeight}
\alias{SunflowerHeight}
\title{Height of a Sunflower}
\source{
Reed, H.S. and Holland, R. H., The Growth of an Annual Plant Helianthus,
\emph{Proceedings of the National Academy of Sciences} (USA), 5, 135-144 (1919).
}
\usage{
data(SunflowerHeight)
}
\description{
A data set containing the height of a sunflower in centimeters
as a function of the day.
}
\details{
A data frame with 13 observations on the following variables.
  \itemize{
    \item{\code{Day}}
    \item{\code{Height}} (in centimeters)
  }
}
\examples{
data(SunflowerHeight)

}
