% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{SunPositionAlaska}
\alias{SunPositionAlaska}
\title{Sun Position in Anchorage, Alaska}
\source{
Altitude angle of the sun in Anchorage, Alaska, each hour from midnight on June 29, 2014
(hour 0) until midnight on June 30, 2014 (hour 24) from “Sun & moon times, Anchorage, Alaska,
U.S.A.” Accessed on July 1, 2014.
}
\usage{
data(SunPositionAlaska)
}
\description{
A data set containing the altitude angle of the sun in Anchorage, Alaska, each hour from
midnight on June 29, 2014 (hour 0) until midnight on June 30, 2014 (hour 24).
}
\details{
A data frame with 25 observations on the following variables.
  \itemize{
    \item{\code{Time}} (hours since midnight on June 29, 2014)
    \item{\code{Position}} (Altitude angle of the sun)
  }
}
\examples{
data(SunPositionAlaska)

}
\references{
http://www.timeanddate.com/astronomy/usa/anchorage
}
