% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{SunriseCancun}
\alias{SunriseCancun}
\title{Sunrise in Cancun, Mexico}
\source{
Cancun, Quintana Roo, Mexico — Sunrise, Sunset, and Daylength, August 2025.
Accessed on August 4, 2025.
}
\usage{
data(SunriseCancun)
}
\description{
A data set containing the time of sunrise in Cancun, Mexico in
minutes after 6 AM on the first day of each month from July 2023 to August 2025.
}
\details{
A data frame with 26 observations on the following variables.
  \itemize{
    \item{\code{Month}} (since July 2023)
    \item{\code{Sunrise}} (minutes after 6 AM)
  }
}
\examples{
data(SunriseCancun)

}
\references{
https://www.timeanddate.com/sun/mexico/cancun
}
