% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{SunriseLA}
\alias{SunriseLA}
\title{Sunrise in Los Angeles, CA}
\source{
Number of minutes after 4 a.m. until sunrise in Los Angeles, adjusted for Daylight Savings Time,
from January 2010 (month 1) through December 2011 (month 24) from “Sunrise and sunset times in
Los Angeles, December 2011.” Accessed on July 14, 2014.
}
\usage{
data(SunriseLA)
}
\description{
A data set containing the number of minutes after 4 a.m. until sunrise in Los Angeles,
California, adjusted for Daylight Savings Time, from January 2010 (month 1) through
December 2011 (month 24).
}
\details{
A data frame with 24 observations on the following variables.
  \itemize{
    \item{\code{Month}} (from January 2010)
    \item{\code{Time}} (minutes after 4 a.m.)
  }
}
\examples{
data(SunriseLA)

}
\references{
http://www.timeanddate.com/sun/usa/los-angeles?month=12&year=2011
}
