% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{SunsetLA}
\alias{SunsetLA}
\title{Sunset in Los Angeles, CA}
\source{
Number of minutes after 4 p.m. until sunset in Los Angeles, California, adjusted for
Daylight Savings Time, from January 2010 (month 1) through December 2013 (month 48)
from “Sunrise and sunset times in Los Angeles, December 2011.” Accessed on July 14, 2014.
}
\usage{
data(SunsetLA)
}
\description{
A data set containing the number of minutes after 4 p.m. until sunset in Los Angeles,
California, adjusted for Daylight Savings time, from January 2010 (month 1) through
December 2013 (month 48).
}
\details{
A data frame with 48 observations on the following variables.
  \itemize{
    \item{\code{Month}} (from January 2010)
    \item{\code{Minutes}} (minutes after 4 p.m.)
  }
}
\examples{
data(SunsetLA)

}
\references{
http://www.timeanddate.com/sun/usa/los-angeles?month=12&year=2011
}
