% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{SwimmingSpeed}
\alias{SwimmingSpeed}
\title{Swimming Speed of Various Animals}
\format{A data frame with 17 observations on the following variables.
  \itemize{
    \item{\code{Length}} {of animal in centimeters}
    \item{\code{Speed}} {swimming speed of animal in centimeters per second}
  }}
\source{
Swimming speed and length of animals from page 152. McMahon, Thomas A., and John Tyler Bonner.
\emph{On Size and Life}. New York: Scientific American Library, 1983.
}
\usage{
data(SwimmingSpeed)
}
\description{
A dataset containing the swimming speed in centimeters per second as a function of the length
in centimeters of various animals
}
\details{
This dataset contains the swimming speed in centimeters per second as a function of the length
in centimeters of various animals.
}
\examples{
data(SwimmingSpeed)

}
