% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{TemperaturesDanville}
\alias{TemperaturesDanville}
\title{Temperatures in Danville, KY}
\format{A data frame with 60 observations on the following variables.
  \itemize{
    \item{\code{Month}} {from January 2006}
    \item{\code{Temperature}} {in degrees Farenheit}
  }}
\source{
Average maximum temperature in Danville, Kentucky at the beginning of each month since January
2006 from “noaa.gov.” Accessed on June 25, 2014
}
\usage{
data(TemperaturesDanville)
}
\description{
A dataset containing the average maximum temperature in Danville, Kentucky at the beginning
of each month since January 2006
}
\details{
This dataset contains the average maximum temperature in Danville, Kentucky at the beginning
of each month since January 2006
}
\examples{
data(TemperaturesDanville)

}
\references{
http://www1.ncdc.noaa.gov/pub/orders/cdo/352625.pdf
}
