% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{TemperaturesLexington}
\alias{TemperaturesLexington}
\title{Average Temperature in Lexington, KY}
\source{
Average maximum temperature in Lexington Kentucky at the beginning of each month
since October 2019 from ``Lexington, KY Weather History. Accessed on January 7, 2025.
}
\usage{
data(TemperaturesLexington)
}
\description{
A data set containing the average maximum temperature in Lexington,
Kentucky at the beginning of each month since October 2019.
}
\details{
A data frame with 61 observations on the following variables.
  \itemize{
    \item{\code{Month}} (since October 2019)
    \item{\code{Temperature}} (in degrees Fahrenheit)
  }
}
\examples{
data(TemperaturesLexington)

}
\references{
https://www.wunderground.com/history/monthly/us/ky/lexington/KLEX/date/2024-10.
}
