% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{TeslaRevenue}
\alias{TeslaRevenue}
\title{Revenue of Tesla}
\source{
Revenue of Tesla from 2008 to 2024 (in millions U.S. dollars).
Accessed on June 18, 2025.
}
\usage{
data(TeslaRevenue)
}
\description{
A data set containing the revenue of Tesla from 2008 to 2024 in
millions of US dollars.
}
\details{
A data frame with 9 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Revenue}}  (in millions of US dollars)
  }
}
\examples{
data(TeslaRevenue)

}
\references{
https://www.statista.com/statistics/272120/revenue-of-tesla/
}
