% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{TherapodLength}
\alias{TherapodLength}
\title{Body length of therapods}
\source{
Therrien, François & Henderson, Donald. (2007). My theropod is bigger than
yours...or not: Estimating body size from skull length in theropods. Journal
of Vertebrate Paleontology. 108-115. doi: 10.1671/0272-4634(2007)27[108:MTIBTY]
2.0.CO;2. Accessed on October 8, 2024.
}
\usage{
data(TherapodLength)
}
\description{
A data set containing the body length in meters of different therapods as a
function of the skull length in meters.
}
\details{
A data frame with 19 observations on the following variables.
  \itemize{
    \item{\code{Skull}} (in meters)
    \item{\code{Body}} (in meters)
  }
}
\examples{
data(TherapodLength)

}
\references{
https://www.researchgate.net/publication/232687833_My_theropod_is_bigger_than_yoursor_not_Estimating_body_size_from_skull_length_in_theropods.
}
