% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{ThreadfinBream}
\alias{ThreadfinBream}
\title{Weight of Threadfin Bream}
\source{
Pauly, D., Some simple methods for the 1983 assessment of tropical fish
stocks.\emph{FAO Fish.Tech.Pap}., (234):52 p.
}
\usage{
data(ThreadfinBream)
}
\description{
A data set containing the weight of threadfin bream measured in
grams as a function of their length in centimeters
}
\details{
A data frame with 16 observations on the following variables.
  \itemize{
    \item{\code{Length}} (in centimeters)
    \item{\code{Weight}}  (in grams)
  }
}
\examples{
data(ThreadfinBream)

}
