% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{ToyotaMonthly}
\alias{ToyotaMonthly}
\title{Toyota Stock Prices}
\source{
Toyota Motors Corporation (TM) stock market value in U.S. dollars from “Yahoo Finance |
TM Historical Prices.” Accessed on August 13, 2016.
}
\usage{
data(ToyotaMonthly)
}
\description{
A data set containing the average monthly stock price of Toyota from 1982 to 1998.
}
\details{
A data frame with 60 observations on the following variables.
  \itemize{
    \item{\code{Month}} (from January 1982)
    \item{\code{Value}} (in US dollars)
  }
}
\examples{
data(ToyotaMonthly)

}
\references{
http://finance.yahoo.com/quote/TM/history?p=TM
}
