% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{Transistors}
\alias{Transistors}
\title{Transistors in an Integrated Circuit}
\source{
Number of transistors in an integrated circuit from ``Moore's Law: The number
of transistors per microprocessor.'' Accessed on May 8, 2024.
}
\usage{
data(Transistors)
}
\description{
A data set containing the number of transistors in an integrated circuit from
1971 to 2021.
}
\details{
A data frame with 39 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Transistors}}
  }
}
\examples{
data(Transistors)

}
\references{
https://ourworldindata.org/grapher/transistors-per-microprocessor.
}
