% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{TwitterUsers}
\alias{TwitterUsers}
\title{Twitter Users}
\format{A data frame with 17 observations on the following variables.
  \itemize{
    \item{\code{Year}} {since 2000}
    \item{\code{Users}} {number of users in millions of people}
  }}
\source{
Number of monthly active Twitter users worldwide from 1st quarter 2010
to 1st quarter 2016 (in millions) from “Twitter : number of monthly active users 2010–2016 |
Statista.” Accessed on June 9, 2016.
}
\usage{
data(TwitterUsers)
}
\description{
A dataset containing the number of monthly active Twitter users worldwide by quarter
(e.g., 10.25 represents April to June 2010) in millions of people.
}
\details{
This dataset contains the number of monthly active Twitter users worldwide by quarter
(e.g., 10.25 represents April to June 2010) in millions of people.
}
\examples{
data(TwitterUsers)

}
\references{
http://www.statista.com/statistics/282087/number-of-monthly-active-twitter-users/
}
