% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{TwitterUsers3}
\alias{TwitterUsers3}
\title{Twitter Users}
\format{A data frame with 24 observations on the following variables.
  \itemize{
    \item{\code{Year}} {since 2000}
    \item{\code{Users}} {number of users in millions of people}
  }}
\source{
Number of monthly active Twitter users worldwide from 1st quarter 2010
to 1st quarter 2016 (in millions) from “Twitter : number of monthly active users 2010–2016 |
Statista.” Accessed on June 9, 2016.
}
\usage{
data(TwitterUsers3)
}
\description{
A dataset containing how many millions of people used Twitter from the first quarter of 2010
to the beginning of 2016.
}
\details{
This dataset contains how many millions of people used Twitter from the first quarter of 2010
to the beginning of 2016
}
\examples{
data(TwitterUsers3)

}
\references{
http://www.statista.com/statistics/282087/number-of-monthly-active-twitter-users/
}
