% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{UKOnlineSales}
\alias{UKOnlineSales}
\title{Online Retail Sales in the United Kingdom}
\source{
Value of online retail sales in the United Kingdom (UK) from 2012 to 2022
(in billion GBP). Accessed on June 24, 2025.
}
\usage{
data(UKOnlineSales)
}
\description{
A data set containing the value of online retail sales in
billions of Great British pounds (GBP) in the United Kingdom from 2012 to 2022.
}
\details{
A data frame with 11 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Sales}} (in billions of GBP)
  }
}
\examples{
data(UKOnlineSales)

}
\references{
https://www.statista.com/statistics/315506/online-retail-sales-in-the-united-kingdom/
}
