% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{UnemploymentRate}
\alias{UnemploymentRate}
\title{United States Unemployment Rate}
\source{
Unemployment rate in the United States from 1990 to 2024. Accessed on
May 28, 2025.
}
\usage{
data(UnemploymentRate)
}
\description{
A data set containing the annual United States unemployment rate
from 1990 to 2024.
}
\details{
A data frame with 35 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Rate}}
  }
}
\examples{
data(UnemploymentRate)

}
\references{
https://www.statista.com/statistics/193290/unemployment-rate-in-the-usa-since-1990/
}
