% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{UrbanPopulation}
\alias{UrbanPopulation}
\title{Urban Population}
\source{
Size of the urban and rural population of the United States from 1960 to
2020 (in millions). Accessed on September 19, 2024.
}
\usage{
data(UrbanPopulation)
}
\description{
A data set containing the urban population of the United States (in millions)
from 1960 to 2020.
}
\details{
A data frame with 61 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Population}} (in millions of people)
  }
}
\examples{
data(UrbanPopulation)

}
\references{
https://www.statista.com/statistics/985183/size-urban-rural-population-us/
}
