% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{USCO2Emissions}
\alias{USCO2Emissions}
\title{US CO2 Emissions}
\source{
United States carbon dioxide emissions in kT annually from 1960 to 2010 according to the
World Bank at “Data | United States.” Accessed on July 10, 2014.
}
\usage{
data(USCO2Emissions)
}
\description{
A data set containing U.S. carbon dioxide emissions in kT (energy) annually from 1980 to 2008
according to the World Bank.
}
\details{
A data frame with 29 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{kT}}
  }
}
\examples{
data(USCO2Emissions)

}
\references{
http://data.worldbank.org/country/united-states
}
