% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{USGDPDebt}
\alias{USGDPDebt}
\title{United States Debt as a Percentage of GDP}
\source{
U.S. Office of Management and Budget and Federal Reserve Bank of St. Louis,
Federal Debt: Total Public Debt as Percent of Gross Domestic Product
[GFDEGDQ188S], retrieved from FRED, Federal Reserve Bank of St. Louis.
Accessed on August 10, 2025.
}
\usage{
data(USGDPDebt)
}
\description{
A data set containing the total public debt of the United
States as a percentage of the US gross domestic product from the first
quarter of 1966 to the first quarter of 2025.
}
\details{
A data frame with 237 observations on the following variables.
  \itemize{
    \item{\code{Quarter}} (since first Quarter of 1966)
    \item{\code{Percent}}
  }
}
\examples{
data(USGDPDebt)

}
\references{
https://fred.stlouisfed.org/series/GFDEGDQ188S
}
