% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{USGrossDomesticProduct}
\alias{USGrossDomesticProduct}
\title{United States Gross Domestic Product}
\source{
U.S. Bureau of Economic Analysis, Gross Domestic Product [GDP], retrieved
from FRED, Federal Reserve Bank of St. Louis. Accessed on August 15, 2025.
}
\usage{
data(USGrossDomesticProduct)
}
\description{
A data set containing the gross domestic product of the United States
in billions of US dollars each quarter from the first quarter of 1966 to the
second quarter of 2025.
}
\details{
A data frame with 238 observations on the following variables.
  \itemize{
    \item{\code{Quarter}} (since first Quarter of 1966)
    \item{\code{GDP}}  (in billions of US dollars)
  }
}
\examples{
data(USGrossDomesticProduct)

}
\references{
https://fred.stlouisfed.org/series/GDP
}
