% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{USInflationRate}
\alias{USInflationRate}
\title{United States Inflation Rate}
\source{
Projected annual inflation rate in the United States from 2010 to 2029.
Accessed on May 31, 2025.
}
\usage{
data(USInflationRate)
}
\description{
A data set containing the annual inflation rate of the United
States from 2010 to 2023.
}
\details{
A data frame with 14 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Rate}}
  }
}
\examples{
data(USInflationRate)

}
\references{
https://www.statista.com/statistics/244983/projected-inflation-rate-in-the-united-states/
}
