% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{USMilitarySpending}
\alias{USMilitarySpending}
\title{United States Military Spending}
\source{
U.S. Military Spending/Defense Budget 1960-2025. Accessed on May 28, 2025.
}
\usage{
data(USMilitarySpending)
}
\description{
A data set containing the annual United States military spending
in billions of US dollars from 1960 to 2022.
}
\details{
A data frame with 35 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Spending}} (in billions of US dollars)
  }
}
\examples{
data(USMilitarySpending)

}
\references{
https://www.macrotrends.net/global-metrics/countries/usa/united-states/military-spending-defense-budget
}
