% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{USNationalDebt}
\alias{USNationalDebt}
\title{United States National Debt}
\source{
U.S. Department of the Treasury. Fiscal Service, Federal Debt: Total Public
Debt [GFDEBTN], retrieved from FRED, Federal Reserve Bank of St. Louis.
Accessed on August 15, 2025.
}
\usage{
data(USNationalDebt)
}
\description{
A data set containing the total United States pulic debt in millions
of US dollars each quarter from the first quarter of 1966 to the first quarter
of 2025.
}
\details{
A data frame with 237 observations on the following variables.
  \itemize{
    \item{\code{Quarter}} (since first Quarter of 1966)
    \item{\code{Debt}}  (in millions of US dollars)
  }
}
\examples{
data(USNationalDebt)

}
\references{
https://fred.stlouisfed.org/series/GFDEBTN
}
