% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{USNuclearTests}
\alias{USNuclearTests}
\title{United States Nuclear Tests}
\source{
United States Nuclear Tests July 1945 through September 1992. U.S. Department
of Energy Nevada Operations Office. DOE/NV—209REV15, December 2000. Accessed
on June 17, 2025.
}
\usage{
data(USNuclearTests)
}
\description{
A data set containing the cumulative number of nuclear tests
conducted by the United States as a function of the year.
}
\details{
A data frame with 48 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Tests}}
  }
}
\examples{
data(USNuclearTests)

}
\references{
https://www.osti.gov/opennet/manhattan-project-history/publications/DOENuclearTests.pdf
}
