% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{USPopulationDecade}
\alias{USPopulationDecade}
\title{United States Population}
\source{
Population of the United States from 1610 to 2020 (in thousands). Accessed
on July 22, 2025.
}
\usage{
data(USPopulationDecade)
}
\description{
A data set containing the population of the United States in
thousands of people each decade.
}
\details{
A data frame with 23 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Population}} (in thousands of people)
  }
}
\examples{
data(USPopulationDecade)

}
\references{
https://www.statista.com/statistics/1067138/population-united-states-historical/
}
