% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{USPowerProduction}
\alias{USPowerProduction}
\title{US Electricity Generation}
\source{
Net electricity generation in the United States from 1950 to 2024
(in terawatt-hours). Accessed on June 18, 2025.
}
\usage{
data(USPowerProduction)
}
\description{
A data set containing the United States net electricity generation
in terawatt-hours from 1950 to 2024.
}
\details{
A data frame with 27 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Power}}  (in terawatt-hours)
  }
}
\examples{
data(USPowerProduction)

}
\references{
https://www.statista.com/statistics/188521/total-us-electricity-net-generation/
}
