% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{USTaxReceipts}
\alias{USTaxReceipts}
\title{United States Tax Receipts}
\source{
Federal government current tax receipts, billions of US dollars, seasonally
adjusted annual rate. Accessed on August 1, 2025
}
\usage{
data(USTaxReceipts)
}
\description{
A data set containing the total amount of United States federal
government receipts in billions of US dollars each quarter from January 1947
to January 2025.
}
\details{
A data frame with 313 observations on the following variables.
  \itemize{
    \item{\code{Quarter}} (since January 1947)
    \item{\code{Receipts}}  (in billions of US dollars)
  }
}
\examples{
data(USTaxReceipts)

}
\references{
https://fred.stlouisfed.org/series/W006RC1Q027SBEA
}
