% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{VehicleRegistrations}
\alias{VehicleRegistrations}
\title{United States Vehicle Registrations}
\source{
Number of motor vehicles registered in the United States from 1990 to 2022
(in 1,000s). Accessed on September 26, 2024.
}
\usage{
data(VehicleRegistrations)
}
\description{
A data set containing the number of motor vehicles registered in the United
States in 1000's from 1990 to 2022.
}
\details{
A data frame with 33 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Vehicles}} (in 1000's)
  }
}
\examples{
data(VehicleRegistrations)

}
\references{
https://www.statista.com/statistics/183505/number-of-vehicles-in-the-united-states-since-1990/
}
