% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{VerizonEmployees}
\alias{VerizonEmployees}
\title{Number of Verizon Employees}
\source{
Number of employees at Verizon from 2007 to 2024. Accessed on August 1, 2025.
}
\usage{
data(VerizonEmployees)
}
\description{
A data set containing the number of employees at Verizon in
thousands from 2007 to 2024.
}
\details{
A data frame with 18 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Employees}}  (in thousands)
  }
}
\examples{
data(VerizonEmployees)

}
\references{
https://www.statista.com/statistics/257304/number-of-employees-at-verizon/
}
