% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{VO2Max}
\alias{VO2Max}
\title{Mammalian maximal metabolic rate}
\source{
Ewald R. Weibel, Leonardo D. Bacigalupe, Beat Schmitt, and Hans Hoppeler.
Allometric scaling of maximal metabolic rate in mammals: muscle aerobic
capacity as determinant factor, Respiratory Physiology & Neurobiology,
Volume 140, Issue 2, 2004, Pages 115-132, ISSN 1569-9048,
https://doi.org/10.1016/j.resp.2004.01.006. Accessed on October 8, 2024.
}
\usage{
data(VO2Max)
}
\description{
A data set containing the maximal metabolic rate in mL/min of various
mammalian species as a function of their mass in kilograms
}
\details{
A data frame with 34 observations on the following variables.
  \itemize{
    \item{\code{Mass}} (in kilograms)
    \item{\code{VO2Max}} (in mL/min)
  }
}
\examples{
data(VO2Max)

}
\references{
https://www.sciencedirect.com/science/article/pii/S1569904804000126.
}
