% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{WaterLevelsEastportMaine}
\alias{WaterLevelsEastportMaine}
\title{Water Levels in Eastport, Maine}
\format{A data frame with 477 observations on the following variables.
  \itemize{
    \item{\code{Hours}} {since 12 a.m. on May 26m, 2016}
    \item{\code{WaterLevel}} {feet above the mean lower water level}
  }}
\source{
Observed water levels in Eastport, Maine from NOAA Tides and Currents. Accessed on June 2, 2016.
}
\usage{
data(WaterLevelsEastportMaine)
}
\description{
A dataset containing the water level in Eastport, Maine from May 26, 2016 to May 27, 2016
measured in feet above the mean lower water level as a function of the time after 12:00 a.m. on
May 26, 2016.
}
\details{
This dataset contains the water level in Eastport, Maine from May 26, 2016 to May 27, 2016
measured in feet above the mean lower water level as a function of the time after 12:00 a.m.
on May 26, 2016.
}
\examples{
data(WaterLevelsEastportMaine)

}
\references{
http://tidesandcurrents.noaa.gov/waterlevels.html?id=8410140&units=standard&bdate=20160526&edate=20160527&timezone=GMT&datum=MLLW&interval=6&action=
}
