% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{WeightChange}
\alias{WeightChange}
\title{Weight Change during Pregnancy}
\source{
Personal data collection
}
\usage{
data(WeightChange)
}
\description{
A data set containing the change in one woman’s weight during a pregnancy from the beginning of
her second trimester until birth.
}
\details{
A data frame with 129 observations on the following variables.
  \itemize{
    \item{\code{Day}} (since beginning of second trimester)
    \item{\code{Weight}} (in pounds)
  }
}
\examples{
data(WeightChange)

}
