% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{WineQuality}
\alias{WineQuality}
\title{Wine Quality of Red Vinho Verde Wines from Portugal}
\source{
Cortez, P., Cerdeira, A., Almeida, F., Matos, T., & Reis, J. (2009). Wine
Quality [Dataset]. UCI Machine Learning Repository. Accessed on May 21, 2025.
}
\usage{
data(WineQuality)
}
\description{
A data set containing data on 1599 red vinho verde wines from
Portugal. For each wine, the alcohol level, the volatile acidity of the wine,
amount of sulphates present in the wine, and a sensory score assigned by a
human expert based on the wines smell is recorded.
}
\details{
A data frame with 1599 observations on the following variables.
  \itemize{
    \item{\code{volatileAcidity}} (in grams per liter)
    \item{\code{sulphates}} (parts per million)
    \item{\code{alcohol}} (percent)
    \item{\code{quality}} (number between 0 and 10)
  }
}
\examples{
data(WineQuality)

}
\references{
https://doi.org/10.24432/C56S3T.
}
