% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{WineSales}
\alias{WineSales}
\title{Sales Volume of Wine in United States}
\source{
Sales volume of wine in the United States from 1999 to 2023 (in million
9-liter cases). Accessed on March 11, 2025.
}
\usage{
data(WineSales)
}
\description{
A data set containing the sales volume of wine in the United
States, measured in millions of 9-liter cases, from 1999 to 2023.
}
\details{
A data frame with 23 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Cases}} (millions of 9-liter cases)
  }
}
\examples{
data(WineSales)

}
\references{
https://www.statista.com/statistics/715474/wine-sales-in-us/
}
