% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{WorldPopulationChange}
\alias{WorldPopulationChange}
\title{World Population Change}
\format{A data frame with 10 observations on the following variables.
  \itemize{
    \item{\code{Year}} {}
    \item{\code{Growth}} {Percent growth of worlds population}
  }}
\source{
World population from the U.S. Census Bureau. Accessed on June 3, 2016.
}
\usage{
data(WorldPopulationChange)
}
\description{
A dataset containing the percent growth of the world’s population as a function of the year
from 1970 to 2015.
}
\details{
This dataset contains the percent growth of the world’s population as a function of the year
from 1970 to 2015.
}
\examples{
data(WorldPopulationChange)

}
\references{
http://www.census.gov/population/international/data/worldpop/table_population.php
}
