% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{XUsers}
\alias{XUsers}
\title{Active X Users Worldwide}
\source{
Number of monthly active X (formerly Twitter) users worldwide from 1st
quarter 2010 to 1st quarter 2019. Accessed on January 15, 2025.
}
\usage{
data(XUsers)
}
\description{
A data set containing the number of active X (formerly Twitter)
users worldwide in millions each quarter from the first quarter of 2020 to
the first quarter of 2019.
}
\details{
A data frame with 37 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Users}} (in millions)
  }
}
\examples{
data(XUsers)

}
\references{
http://www.statista.com/statistics/282087/number-of-monthly-active-twitter-users/
}
