% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{APCalculus}
\alias{APCalculus}
\title{Number of Students taking the AP Calculus Exam}
\format{A data frame with 61 observations on the following variables.
  \itemize{
    \item{\code{Year}} {year}
    \item{\code{Exams}} {number of AP exams taken in associated year}
  }}
\source{
Personal correspondence with Stephen M. Kokoska.
}
\usage{
data(APCalculus)
}
\description{
A dataset containing the number of students who have taken the AP Calculus exam from 1955 to 2015.
}
\details{
The number of AP exams taken in the given year, with data from 1955 to 2015.
}
\examples{
data(APCalculus)

}
