% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{EbolaSierraLeone}
\alias{EbolaSierraLeone}
\title{Ebola Cases in Sierra Leone, Africa}
\format{A data frame with 110 observations on the following variables.
  \itemize{
    \item{\code{Day}} {since May 1, 2014}
    \item{\code{Cases}} {cummulative number of ebola cases}
  }}
\source{
Ebola Outbreak in West Africa from Centers for Disease Control and Preven- tion. Accessed on
June 2, 2016.
}
\usage{
data(EbolaSierraLeone)
}
\description{
A dataset containing the cummulative number of ebola cases in Sierra Leone, Africa from
May 1, 2014 to December 16, 2015
}
\details{
This dataset contains the cummulative number of ebola cases in Sierra Leone, Africa from
May 1, 2014 to December 16, 2015
}
\examples{
data(EbolaSierraLeone)

}
\references{
http://www.cdc.gov/vhf/ebola/outbreaks/2014-west-africa/cumulative-cases-graphs.html
}
