% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{ElectricBill}
\alias{ElectricBill}
\title{Electric Bill}
\format{A data frame with 37 observations on the following variables.
  \itemize{
    \item{\code{Month}} {since January 2000}
    \item{\code{ElecBill}} {price of ElecBill in US dollars}
  }}
\source{
Subset of Utilities in the \code{mosaic} package.
}
\usage{
data(ElectricBill)
}
\description{
A dataset containing the electric bill (in US dollars) of a single-family home in Minnesota from
2000-2003
}
\details{
This dataset contains the the electric bill (in US dollars) of a single-family home in Minnesota from
2000-2003 as a function of the month.
}
\examples{
data(ElectricBill)

}
