% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{ElectronicMailOrderSales}
\alias{ElectronicMailOrderSales}
\title{US Electronic and Mail Order Sales}
\source{
Total U.S. electronic and mail-order shopping sales in millions of
dollars each year from 1999 to 2012 from “Monthly & Annual Retail Trade,
Main Page — US Census Bureau.” Accessed on July 16, 2014.
}
\usage{
data(ElectronicMailOrderSales)
}
\description{
A data set containing the he total U.S. electronic and mail-order
shopping sales in billions of dollars each year from 1999 to 2012.
}
\details{
A data frame with 14 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Sales}} (in billions of US dollars)
  }
}
\examples{
data(ElectronicMailOrderSales)

}
