% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{GenderRatio2}
\alias{GenderRatio2}
\title{Gender Ratio in World Population}
\source{
The global gender ratio based on the number of males per 100 females by year from
“World Population Prospects, the 2012 Revision” by the United Nations Department of
Economic and Social Affairs. Accessed on June 25, 2014.
}
\usage{
data(GenderRatio2)
}
\description{
A data set containing the number of males per 100 females in the word population.
}
\details{
A data frame with 14 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Ratio}} (number of males per 100 females)
  }
}
\examples{
data(GenderRatio2)

}
\references{
http://esa.un.org/unpd/wpp/Excel-Data/population.htm
}
