% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{McDBurgers2}
\alias{McDBurgers2}
\title{Burgers Sold by McDonalds}
\source{
The total number of burgers sold by McDonald’s in billions as of each year from “Over How
Many Billion Served.” Accessed on July 3, 2014.
}
\usage{
data(McDBurgers2)
}
\description{
A data set containing the number of burgers (in billions) sold by McDonald’s since 1955.
}
\details{
A data frame with 11 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Burgers}} (in billions)
  }
}
\examples{
data(McDBurgers2)

}
\references{
http://overhowmanybillionserved.blogspot.com/
}
