% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{NaturalGasConsumption}
\alias{NaturalGasConsumption}
\title{US Natural Gas Consumption}
\format{A data frame with 21 observations on the following variables.
  \itemize{
    \item{\code{Year}} {}
    \item{\code{CubicFeet}} {millions of cubic feet of natural gas consumed}
  }}
\source{
Natural Gas from U.S. Energy Information Administration. Accessed on June 9, 2016.
}
\usage{
data(NaturalGasConsumption)
}
\description{
A subset of the dataset containing the total number of millions of cubic feet of natural gas consumed
in the United States from 1950 to 1970.
}
\details{
This dataset contains the total number of millions of cubic feet of natural gas
consumed in the United States from 1950 to 1970.
}
\examples{
data(NaturalGasConsumption)

}
\references{
http://www.eia.gov/dnav/ng/hist/n9140us2a.htm
}
