% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{NaturalGasConsumption2}
\alias{NaturalGasConsumption2}
\title{US Natural Gas Consumption}
\source{
Natural Gas from U.S. Energy Information Administration. Accessed on June 9, 2016.
}
\usage{
data(NaturalGasConsumption2)
}
\description{
A data set containing the total number of millions of cubic feet of natural gas consumed
in the United States from 1950 to 1970.
}
\details{
A data frame with 67 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{CubicFeet}} (in millions of cubic feet)
  }
}
\examples{
data(NaturalGasConsumption2)

}
\references{
http://www.eia.gov/dnav/ng/hist/n9140us2a.htm
}
