% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{NetherlandsPopulation}
\alias{NetherlandsPopulation}
\title{Population of the Netherlands}
\format{A data frame with 21 observations on the following variables.
  \itemize{
    \item{\code{Year}} {}
    \item{\code{Population}} {of Netherlands (in millions of people)}
  }}
\source{
Population of the Netherlands from “The NETHERLANDS : country populations.”
Accessed on July 11, 2014.
}
\usage{
data(NetherlandsPopulation)
}
\description{
A dataset containing the population of the Netherlands measured in millions of people
at the beginning of each decade since 1700.
}
\details{
This dataset contains the population of the Netherlands measured in millions of people
at the beginning of each decade since 1700.
}
\examples{
data(NetherlandsPopulation)

}
\references{
http://www.populstat.info/Europe/netherlc.htm
}
