% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{RunningSpeed}
\alias{RunningSpeed}
\title{Running Speed of Animals}
\source{
Running speed and length of animals from page 152. McMahon, Thomas A., and John Tyler Bonner.
\emph{On Size and Life}. New York: Scientific American Library, 1983.
}
\usage{
data(RunningSpeed)
}
\description{
A data set containing the running speed in centimeters per second as a function of the
length in centimeters of various animals.
}
\details{
A data frame with 12 observations on the following variables.
  \itemize{
    \item{\code{Length}} (in centimeters)
    \item{\code{Speed}} (in centimeters per second)
  }
}
\examples{
data(RunningSpeed)

}
