% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{SunsetGreenwich}
\alias{SunsetGreenwich}
\title{Sunset in Greenwich, England}
\source{
Number of minutes after 3 p.m. until sunset in Greenwich, England since January 2010 from
“Sunrise and sunset times in Greenwich Borough.” Accessed on July 14, 2014.
}
\usage{
data(SunsetGreenwich)
}
\description{
A data set containing the number of minutes after 3 p.m. until sunset at Greenwich, England
since January 2010.
}
\details{
A data frame with 25 observations on the following variables.
  \itemize{
    \item{\code{Month}} (from January 2010)
    \item{\code{Minutes}} (minutes after 3 p.m.)
  }
}
\examples{
data(SunsetGreenwich)

}
\references{
http://www.timeanddate.com/sun/uk/greenwich-city
}
