% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{USRetailTax}
\alias{USRetailTax}
\title{US Retail Tax}
\source{
Annual total retail sales taxes collected in the United States in each year from
“Monthly & Annual Retail Trade, Main Page — U.S. Census Bureau.” Accessed on July 14, 2014.
}
\usage{
data(USRetailTax)
}
\description{
A data set containing U.S. retail tax in millions of dollars each year from 2005 through 2011.
}
\details{
A data frame with 7 observations on the following variables.
  \itemize{
    \item{\code{Year}}
    \item{\code{Tax}} (in millions of dollars)
  }
}
\examples{
data(USRetailTax)

}
\references{
http://www.census.gov/retail/
}
