% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{World Population}
\format{A data frame with 14 observations on the following variables.
  \itemize{
    \item{\code{Year}} {}
    \item{\code{People}} {world population in billions of people}
  }}
\source{
Estimated world population data from 1950 to 2015 from “International Programs — Total
Mid-Year Population for the World: 1950–2050 — U.S. Census Bureau.” Accessed on June 16, 2014.
}
\usage{
data(WorldPopulation)
}
\description{
A dataset containing the total midyear population for the world from 1950 to 2015.
}
\details{
This dataset contains the total midyear population for the world from 1950 to 2015.
}
\examples{
data(WorldPopulation)

}
\references{
http://www.census.gov/population/international/data/worldpop/table_population.php
}
